Add support to multiple user shares in ExpenseByShare
When creating an expense by share Splitwise API expects a list of
(userID, paid share, owed share) as in users__{index}__{property}*.

Now clients pass the Expense and a list of UserShare. Internally, we
handle the list of UserShares and add to a dynamic struct that
represents the ExpenseByShare that will be sent as the request body.